Re-implement the dictionary comparer
The dictionary comparer was not working properly behaving properly
previously. It was dependent on implementations of GetHashCode and
Equals for objects used as keys in the dictionaries. However, this did
not supportusing instances as keys in dictionaries in combination with
cloning (which is a very common scenario when testing serialization
frameworks). This rewritten version recursively uses this library to
check equality for keys and values. It also uses a heuristic to
drastically reduce comparison times in the normal cases where ordering
is equal, however, a fallback is provided if that is not the case.